put the jwt test directly in oauth20profilecontrollertest
